refactor(worker): separate routers from durable objects
Split Gallery HTTP policy from GalleryDO storage, Auth routing from AuthDO identity/provider storage, and Agent public protocol helpers from AgentSessionDO state. Keep the original module facades and worker/index.ts exports so Wrangler binding class names and migration tags remain unchanged.

Hot files now have explicit ownership: gallery router 24.5 kB and DO 38.9 kB; Agent runtime 17.9 kB and DO 29.6 kB; Auth facade 1.2 kB and DO 25.9 kB.

Validation: pnpm ci:static; complete worker unit suite (76); Agent browser E2E (1); focused Auth/publish/workspace browser E2E (3); git diff --check. Wrangler dry-run was unavailable because this workspace does not install the wrangler CLI; TypeScript verified all worker class exports.

No test files changed because existing Worker and browser contracts exercise the unchanged facades and Durable Object behavior.

Unrelated user-owned probe-conflicts.mjs remained untracked and untouched.

Test-Impact: no-test-change — behavior-preserving module split covered by 76 Worker tests and 4 browser scenarios.
